requirements.txt
fastapi
uvicorn[standard]
python-multipart
boto3
httpx
python-jose[cryptography]
